app/requirements.txt

fastapi>=0.104.0
uvicorn>=0.23.2

motor[srv]>=3.3.1


pydantic>=2.4.2
pydantic[email]>=2.4.2
passlib[bcrypt]>=1.7.4
PyJWT>=2.8.0

opencv-python-headless>=4.8.1.78
tensorflow>=2.14.0
numpy>=1.26.1


cloudinary>=1.36.0


adafruit_io>=2.7.1


python-dotenv>=1.0.0